Add vrsc/zec networks to zcash_protocol for constants
