+++ components/zcash_protocol/src/constants/vrsc.rs
@@ -1,0 +1,3 @@
+pub mod mainnet;
+pub mod testnet;
+pub mod regtest;
